backend/pyproject.toml
[project]
name = "car-charging-reservation-backend"
version = "0.1.0"
description = "Car Charging Reservation System - ATDD Walking Skeleton"
requires-python = ">=3.12"
dependencies = [
    "certifi>=2025.10.5",
    "fastapi>=0.104.0",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "uvicorn[standard]>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "httpx>=0.25.0",
    "ruff>=0.1.0",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "httpx>=0.25.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 88
fix = true

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v"
